requirements.txt
streamlit
python-dotenv
azure-cognitiveservices-speech
openai